Bottom screen changes, early menu, all of the bunch.

Bottom screen works correct but things still need to be done. 

Overall, things still need to be done for this to achieve near completion. Working on that and will merge in the coming days.

Ignore it saying i edited 30000 lines, i did not lmao

+++ src/utils.h
@@ -18,7 +18,7 @@
 
 // The "typeof((typeof(n))0" is used to remove the "const" from the typeof
 
-#define repeat(n, it) for (size_t it = 0; it < n; it++)
+#define repeat(n, it) for (int it = 0; it < (int)(n); it++)
 
 #define require(condition) \
     do { \
